refactor: moved OneOf- and TypeSectionToApie into separate units
