+++ drafter.gyp
@@ -255,10 +255,18 @@
         "src/NodeInfo.h",
         "src/Serialize.h",
         "src/Serialize.cc",
+        "src/SerializeKey.h",
+        "src/SerializeKey.cc",
         "src/SerializeResult.h",
         "src/SerializeResult.cc",
         "src/RefractAPI.h",
         "src/RefractAPI.cc",
+        "src/MsonTypeSectionToApie.h",
+        "src/MsonTypeSectionToApie.cc",
+        "src/MsonMemberToApie.h",
+        "src/MsonMemberToApie.cc",
+        "src/MsonOneOfSectionToApie.h",
+        "src/MsonOneOfSectionToApie.cc",
         "src/RefractDataStructure.h",
         "src/RefractDataStructure.cc",
         "src/RefractSourceMap.h",
