+++ src/MsonMemberToApie.cc
@@ -1,0 +1,9 @@
+//
+//  MsonMemberToApie.cc
+//  drafter
+//
+//  Created by Thomas Jandecka on 12/07/19.
+//  Copyright (c) 2019 Apiary Inc. All rights reserved.
+//
+
+#include "MsonMemberToApie.h"
